Admin List Map refactor start
