+++ src/pages/countries/components/navigation/index.jsx
@@ -4,18 +4,18 @@ import { NavLink } from 'react-router-dom';
 const TabNavigation = () => {
   return (
     <div className='flex items-center w-[100%] h-[10%] p-4'>
-      <div className='flex w-full items-center gap-8'>
+      <div className='flex w-full items-center gap-8 child:pb-2'>
         <NavLink
           to='/countries'
           className={({ isActive }) =>
-            isActive ? 'text-red-300' : 'text-afexgreen'
+            isActive ? 'border-b-2 border-b-afexgreen' : undefined
           }>
           Heat Map
         </NavLink>
         <NavLink
           to='/countries/list'
           className={({ isActive }) =>
-            isActive ? 'text-red-300' : 'text-afexgreen'
+            isActive ? 'border-b-2 border-b-afexgreen' : undefined
           }>
           Countries List
         </NavLink>
